Specify columns for tables
Very specific to Credit Cooperatif releve from a particular date.

This settings resolve the problem of last table on last page
containing few operations and not detected as a "bank operations" table.
To avoid to treat junk, after line containing "NOUVEAU SOLDE ..." we consider
that we are not in a "bank operations" table anymore.